requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# Database
sqlite3
asyncpg>=0.29.0  # For PostgreSQL if needed
redis>=5.0.0

# AI Providers
openai>=1.5.0
anthropic>=0.8.0
requests>=2.31.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Payment Processing
stripe>=7.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Monitoring & Analytics
prometheus-client>=0.19.0
structlog>=23.2.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# API Documentation
python-multipart>=0.0.6

# Caching
redis>=5.0.0

# Background Tasks
celery>=5.3.0

# Email
emails>=0.6

# File Processing
aiofiles>=23.2.1

# Rate Limiting
slowapi>=0.1.9

# Utilities
click>=8.1.7
rich>=13.7.0